+++ jolt-core/src/subprotocols/mod.rs
@@ -2,4 +2,5 @@
 
 pub mod grand_product;
 pub mod grand_product_quarks;
+pub mod sparse_grand_product;
 pub mod sumcheck;
